requirements.txt
feedparser==6.0.11
jinja2==3.1.4
pyyaml==6.0.1